tests/wpt/web-platform-tests/workers/interfaces/WorkerGlobalScope/self.html

<!doctype html>
<title>self</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
async_test(function() {
  var worker = new Worker('self.js');
  worker.onmessage = this.step_func(function(e) {
    for (var i = 0; i < e.data.length; ++i) {
      assert_true(e.data[i][0], e.data[i][1]);
    }
    this.done();
  });
});
</script>
